Flatten SslStream by removing SslStreamInternal SslState (#35109)
* Still moving sslState

* Initial SslState removed, should build

* Fix tests

* Moved first methods from SslStreamInternal

* moved a few more messages

* Moved more SslStreamInternal into SslStream

* Move half of SslStreamInternal into SslStream now

* Almost all moved other than shutdown

* Removed SslStreamInternal completely

* Final cleanup

* Clean Fakes

* Formatting

* Remove useless properties, remove dispose internal

* Remove HandshakeComplete and IsShutdown, removed methods from the fakes that match

* Added comment on Fake around removing build warnings
