+++ src/System.Net.Security/tests/UnitTests/System.Net.Security.Unit.Tests.csproj
@@ -1,4 +1,4 @@
-<Project Sdk="Microsoft.NET.Sdk">
+﻿<Project Sdk="Microsoft.NET.Sdk">
   <PropertyGroup>
     <ProjectGuid>{0D174EA9-9E61-4519-8D31-7BD2331A1982}</ProjectGuid>
     <AllowUnsafeBlocks>true</AllowUnsafeBlocks>
@@ -20,7 +20,7 @@
     <Compile Include="System\Security\Authentication\InvalidCredentialExceptionTest.cs" />
     <Compile Include="TlsAlertsMatchWindowsInterop.cs" />
     <!-- Fakes -->
-    <Compile Include="Fakes\FakeSslState.cs" />
+    <Compile Include="Fakes\FakeSslStream.Implementation.cs" />
     <Compile Include="Fakes\FakeAuthenticatedStream.cs" />
     <Compile Include="Fakes\FakeLazyAsyncResult.cs" />
     <!-- Common test files -->
@@ -45,6 +45,12 @@
     <Compile Include="..\..\src\System\Net\Security\SslApplicationProtocol.cs">
       <Link>ProductionCode\System\Net\Security\SslApplicationProtocol.cs</Link>
     </Compile>
+    <Compile Include="..\..\src\System\Net\Security\SslConnectionInfo.cs">
+      <Link>ProductionCode\System\Net\Security\SslConnectionInfo.cs</Link>
+    </Compile>
+    <Compile Include="..\..\src\System\Net\Security\SslStream.Implementation.Adapters.cs">
+      <Link>ProductionCode\System\Net\Security\SslStream.Implementation.Adapters.cs</Link>
+    </Compile>
     <Compile Include="..\..\src\System\Net\SslStreamContext.cs">
       <Link>ProductionCode\System\Net\SslStreamContext.cs</Link>
     </Compile>
@@ -59,6 +65,9 @@
     </Compile>
     <Compile Include="$(CommonPath)\Interop\Windows\SChannel\Interop.Alerts.cs">
       <Link>Common\Interop\Windows\SChannel\Interop.Alerts.cs</Link>
+    </Compile>
+    <Compile Include="$(CommonPath)\CoreLib\System\Threading\Tasks\TaskToApm.cs">
+      <Link>Common\CoreLib\System\Threading\Tasks\TaskToApm.cs</Link>
     </Compile>
     <!-- Logging -->
     <Compile Include="$(CommonPath)\System\Net\Logging\NetEventSource.Common.cs">
